Moved `WizardPage`s to separate package.
