--- Core/src/main/java/de/uka/ipd/sdq/beagle/gui/MachineSelectionWizardPage.java
@@ -1,8 +1,0 @@
-package de.uka.ipd.sdq.beagle.gui;
-
-public class MachineSelectionWizardPage extends WizardPage {
-
-	public MachineSelectionWizardPage(UserConfiguration userConfiguration) {
-
-	}
-}
