--- Core/src/main/java/de/uka/ipd/sdq/beagle/gui/TimeoutWizardPage.java
@@ -1,8 +1,0 @@
-package de.uka.ipd.sdq.beagle.gui;
-
-public class TimeoutWizardPage extends WizardPage {
-
-	public TimeoutWizardPage(UserConfiguration userConfiguration) {
-
-	}
-}
