+++ Core/src/main/java/de/uka/ipd/sdq/beagle/gui/wizardpages/MachineSelectionWizardPage.java
@@ -1,0 +1,11 @@
+package de.uka.ipd.sdq.beagle.gui.wizardpages;
+
+import de.uka.ipd.sdq.beagle.gui.UserConfiguration;
+import de.uka.ipd.sdq.beagle.gui.WizardPage;
+
+public class MachineSelectionWizardPage extends WizardPage {
+
+	public MachineSelectionWizardPage(UserConfiguration userConfiguration) {
+
+	}
+}
