+++ Core/src/main/java/de/uka/ipd/sdq/beagle/gui/wizardpages/NetworkConfigurationWizardPage.java
@@ -1,4 +1,7 @@
-package de.uka.ipd.sdq.beagle.gui;
+package de.uka.ipd.sdq.beagle.gui.wizardpages;
+
+import de.uka.ipd.sdq.beagle.gui.UserConfiguration;
+import de.uka.ipd.sdq.beagle.gui.WizardPage;
 
 public class NetworkConfigurationWizardPage extends WizardPage {
