+++ Core/src/main/java/de/uka/ipd/sdq/beagle/gui/wizardpages/TimeoutWizardPage.java
@@ -1,0 +1,11 @@
+package de.uka.ipd.sdq.beagle.gui.wizardpages;
+
+import de.uka.ipd.sdq.beagle.gui.UserConfiguration;
+import de.uka.ipd.sdq.beagle.gui.WizardPage;
+
+public class TimeoutWizardPage extends WizardPage {
+
+	public TimeoutWizardPage(UserConfiguration userConfiguration) {
+
+	}
+}
