+++ Core/src/main/java/de/uka/ipd/sdq/beagle/gui/wizardpages/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all {@link WizardPage}s of the Graphicac User Interface (GUI).
+ */
+package de.uka.ipd.sdq.beagle.gui.wizardpages;
